requirements.txt
fastapi
uvicorn
anthropic
pypdf
scikit-learn
numpy
python-dotenv
